+++ activerecord/lib/active_record/locking/optimistic.rb
@@ -85,10 +85,14 @@ module ActiveRecord
 
             self[locking_column] += 1
 
-            affected_rows = self.class._update_record(
-              attributes_with_values(attribute_names),
-              @primary_key => id_in_database,
-              locking_column => previous_lock_value
+            # When `lock_version` is read as `0`, it may actually be `NULL` in the DB.
+            possible_previous_lock_value = previous_lock_value.to_i == 0 ? [nil, 0] : previous_lock_value
+
+            affected_rows = self.class.unscoped.where(
+              locking_column => possible_previous_lock_value,
+              self.class.primary_key => id_in_database
+            ).update_all(
+              attributes_with_values(attribute_names)
             )
 
             if affected_rows != 1